refactor: event_logger.py の Payload クラス群を log_payload.py に分離
handlers/ 配下の大半のファイルは EventLogger 本体ではなく Payload
サブクラスのみに依存していたため、ログのスキーマ定義（Payload群）と
ロガーの管理ロジック（EventLog/EventLogger）を別モジュールに分割し、
依存関係を明確にした。
