+++ src/jpoke/core/turn_controller.py
@@ -9,7 +9,7 @@ if TYPE_CHECKING:
     from jpoke.core import Battle, Player
 
 from jpoke.core import EventContext
-from jpoke.core.event_logger import TerastalPayload
+from jpoke.core.log_payload import TerastalPayload
 from jpoke.enums import Event, Command, Interrupt, LogCode
 from jpoke.utils import fast_copy
