+++ src/jpoke/handlers/item.py
@@ -14,7 +14,7 @@ from jpoke.types import RoleSpec, Stat, Type, MoveCategory, \
 from jpoke.utils.math import apply_fixed_modifier, round_half_down
 from jpoke.enums import Interrupt, LogCode, Command
 from jpoke.core import HandlerReturn, Handler
-from jpoke.core.event_logger import ItemPayload, StatChangePayload
+from jpoke.core.log_payload import ItemPayload, StatChangePayload
 from jpoke.data import TYPE_MODIFIER
 from jpoke.data.megaevol import MEGA_STONES
 from jpoke.data.pokedex import POKEDEX
